Modulo Login corregido

+++ src/services/auth.service.js
@@ -1,16 +1,25 @@
 import http from "../utils/http-common";
-//import authHeader from "./auth-header";
+import authHeader from "./auth-header";
 
-function signin(email, password) {
+function signin(email, password, keepLogged) {
     return http.post("/auth/signin", { email, password })
         .then(response => {
-            if (response.status === 200) 
-                localStorage.setItem("user", JSON.stringify(response.data));
+            if (response.status === 200) {
+                if(keepLogged) {
+                    localStorage.setItem("user", JSON.stringify(response.data));
+                } else {
+                    sessionStorage.setItem("user", JSON.stringify(response.data));
+                }
+            }
         })
         .catch(err => {
             console.log(err.message);
         })
 };
+
+function isLogged() {
+    return http.get("/auth/check", { headers: authHeader() });
+}
 
 function signup(data) {
     return http.post(`/auth/signup`, data);
@@ -28,7 +37,8 @@ let methods = {
     signin,
     signup,
     logout,
-    getCurrentUser
+    getCurrentUser,
+    isLogged
 };
 
 export default methods;
